requirements.txt
streamlit>=1.24.0
pandas>=1.5.0
neo4j>=5.8.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
requests>=2.31.0
beautifulsoup4>=4.12.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
pdfplumber>=0.9.0
python-docx>=0.8.11
PyPDF2>=3.0.0
reportlab>=4.0.0
ipykernel
